feat: add sks information on each jadwal in daftar jadwal
